refactor: minor refactor to main executable
makes subcommand required and moves helper functions to separate files
for improved readability and maintenance
